1. Adjusted access modifiers & virtual modifiers on OutputChannel & company to allow for Wrappers 2. Created LidgrenClientChannelWrapper for decorators 3. Created SharedLidgrenClientChannel decorator 4. Created DurableLidgrenClientChannel decorator 5. Wrote integration tests for 2, 3, and 4
Note that integration tests do not all pass at once. This is due to the static class TransportLayerResolver, and I don't think I can fix this without gutting TransportLayerResolver to be determined via dependency injection. That would be both a terrible mess and a scope creep I don't wish to deal with.